requirements.txt
# --------- pytorch --------- #
# torch>=1.8.1
# torchvision>=0.9.1
pytorch-lightning
timm

# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0

# --------- loggers --------- #
comet-ml
wandb

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
scikit-learn    # used in some callbacks
seaborn         # used in some callbacks
jupyterlab      # better jupyter notebooks
pudb            # debugger
